bug fixes, improvements, code refactoring
issue #94 should be fixed
issue #95 should be mitigated now
Some improvement has been done for #93
