+++ BappDescription.html
@@ -16,8 +16,6 @@
 	<li>Scrollable main tool tabs</li>
 	<li>Scrollable Repeater and Intruder tabs</li>
 	<li>Taking screenshot of repeater or intruder tabs</li>
-	<li>Filtering using title names or custom styles (disabled since version 2022.6 for compatibility reasons)</li>
-	<li>Searching only in filtered items (disabled since version 2022.6 for compatibility reasons)</li>
 	<li>Trimming long titles into 100 characters</li>
 	<li>Show previously chosen titles for a tab</li>
 	<li>Several keyboard shortcuts to make the tab navigation easier</li>
