+++ src/main/java/burp/BurpExtension.java
@@ -1,0 +1,33 @@
+// Burp Suite Sharpener
+// Released as open source by MDSec - https://www.mdsec.co.uk
+// Developed by Soroush Dalili (@irsdl)
+// Project link: https://github.com/mdsecresearch/BurpSuiteSharpener
+// Released under AGPL see LICENSE for more information
+
+package burp;
+
+/*
+ * Burp Suite Sharpener
+ *
+ *
+ * Mainly Developed by:
+ *     Soroush Dalili (@irsdl)
+ *
+ * Some used modules have also been developed by:
+ *     Corey Arthur (@CoreyD97)
+ *
+ * Project link: https://github.com/irsdl/BurpSuiteSharpener
+ *
+ * Released under A GPL v3.0 see LICENSE for more information
+ *
+ * */
+
+import com.irsdl.burp.sharpener.SharpenerBurpExtension;
+
+public class BurpExtension extends SharpenerBurpExtension {
+
+    public static void main(String[] args) {
+        //System.out.println("You have built the extension. You shall play with the jar file now!");
+    }
+
+}
