+++ src/main/java/com/irsdl/burp/sharpener/uiControllers/mainTabs/MainTabsSettings.java
@@ -59,7 +59,7 @@ public class MainTabsSettings extends StandardSettings {
     }
 
     @Override
-    public synchronized void loadSettings() {
+    public void loadSettings() {
         sharedParameters.printDebugMessage("loadSettings");
         MainTabsStyleHandler.resetMainTabsStylesFromSettings(sharedParameters);
         // Adding listeners to Main Tool Tabs
